requirements.txt
fastapi
jinja2
python-dotenv
uvicorn
requests
sqlalchemy
sqlalchemy_utils
pymysql
gitpython